requirements.txt
streamlit==1.32.0
google-generativeai==0.5.4
networkx==3.2.1
faiss-cpu==1.8.0
sentence-transformers==2.7.0
requests==2.31.0
python-dotenv==1.0.1
PyMuPDF==1.24.1
pandas==2.2.1
plotly==5.20.0
streamlit-agraph==0.0.45
py2cytoscape==0.7.1
pytest==8.1.1
black==24.3.0